[tests] -Fix: uniq_name: wrong array index integer width for printf %d
git-svn-id: svn://repo.hu/pcb-rnd/trunk@37253 1d61de99-560f-4dba-8957-1e505d42358b

+++ tests/uniq_name/tester.c
@@ -55,7 +55,7 @@ int main()
 			if (e->value == NULL)
 				printf("(nil)'\n");
 			else
-				printf("#%d'\n", ((int *)e->value) - values);
+				printf("#%d'\n", (int)(((int *)e->value) - values));
 		}
 	}
